+++ phtree/common/b_plus_tree_hash_map.h
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-#ifndef PHTREE_COMMON_B_PLUS_TREE_H
-#define PHTREE_COMMON_B_PLUS_TREE_H
+#ifndef PHTREE_COMMON_B_PLUS_TREE_HASH_MAP_H
+#define PHTREE_COMMON_B_PLUS_TREE_HASH_MAP_H
 
 #include "bits.h"
 #include <cassert>
 #include <tuple>
 #include <vector>
+#include <unordered_map>
 
 /*
  * PLEASE do not include this file directly, it is included via common.h.
@@ -31,12 +32,27 @@
 namespace improbable::phtree {
 
 /*
- * The b_plus_tree_map is a B+tree implementation that uses a hierarchy of horizontally
+ * The b_plus_tree_hash_map is a B+tree implementation that uses a hierarchy of horizontally
  * connected nodes for fast traversal through all entries.
  *
- * Behavior:
- * This is a key-value map. Keys are unique, so for every key there is at most one entry.
+ * Behavior
+ * ========
+ * This is a hash set/map. It behaves just like std::unordered_set / std::unordered_map, minus
+ * some API functions.
+ * The set/map is ordered by their hash.  Entries with identical hash have no specific ordering
+ * but the order is stable with respect to insertion/removal of other entries.
  *
+ *
+ * Rationale
+ * =========
+ * This implementations is optimized for small entry count (for the multi-map PH-tree we
+ * expect small numbers of entries that actually have identical positions), however it should
+ * scale well with large entry counts (it is a tree, so there is no need for rehashing).
+ * Benchmarks show 10%-20% performance improvements for relocate() when using this custom set/map.
+ *
+ *
+ * Internals
+ * =========
  * The individual nodes have at most M entries.
  * The tree has O(log n) lookup and O(M log n) insertion/removal time complexity,
  * space complexity is O(n).
@@ -59,12 +75,9 @@ namespace improbable::phtree {
  *   - There is no guarantee that siblings are on the same depth of the tree.
  * - The tree is not balanced
  *
- * TODO since this is a "map" (with 1:1 mapping of key:value), we could optimize splitting and
- *   merging by trying to reduce `dead space`
- *   (space between key1 and key2 that exceeds (key2 - key1)).
  */
-template <typename T, std::uint64_t COUNT_MAX>
-class b_plus_tree_map {
+template <typename T, typename HashT = std::hash<T>, typename PredT = std::equal_to<T>>
+class b_plus_tree_hash_set {
     class bpt_node_base;
     template <typename ThisT, typename EntryT>
     class bpt_node_data;
@@ -72,32 +85,33 @@ class b_plus_tree_map {
     class bpt_node_inner;
     class bpt_iterator;
 
-    using key_t = std::uint64_t;
+    using hash_t = std::uint32_t;
 
-    using bpt_entry_inner = std::pair<key_t, bpt_node_base*>;
-    using bpt_entry_leaf = std::pair<key_t, T>;
+    using bpt_entry_inner = std::pair<hash_t, bpt_node_base*>;
+    using bpt_entry_leaf = std::pair<hash_t, T>;
 
     using IterT = bpt_iterator;
     using NodeT = bpt_node_base;
     using NLeafT = bpt_node_leaf;
     using NInnerT = bpt_node_inner;
     using LeafIteratorT = decltype(std::vector<bpt_entry_leaf>().begin());
-    using TreeT = b_plus_tree_map<T, COUNT_MAX>;
+    using TreeT = b_plus_tree_hash_set<T, HashT, PredT>;
 
   public:
-    explicit b_plus_tree_map() : root_{new NLeafT(nullptr, nullptr, nullptr)}, size_{0} {};
+    explicit b_plus_tree_hash_set() : root_{new NLeafT(nullptr, nullptr, nullptr)}, size_{0} {};
 
-    b_plus_tree_map(const b_plus_tree_map& other) : size_{other.size_} {
+    b_plus_tree_hash_set(const b_plus_tree_hash_set& other) : size_{other.size_} {
         root_ = other.root_->is_leaf() ? new NLeafT(*other.root_->as_leaf())
                                        : new NInnerT(*other.root_->as_inner());
     }
 
-    b_plus_tree_map(b_plus_tree_map&& other) noexcept : root_{other.root_}, size_{other.size_} {
+    b_plus_tree_hash_set(b_plus_tree_hash_set&& other) noexcept
+    : root_{other.root_}, size_{other.size_} {
         other.root_ = nullptr;
         other.size_ = 0;
     }
 
-    b_plus_tree_map& operator=(const b_plus_tree_map& other) {
+    b_plus_tree_hash_set& operator=(const b_plus_tree_hash_set& other) {
         assert(this != &other);
         delete root_;
         root_ = other.root_->is_leaf() ? new NLeafT(*other.root_->as_leaf())
@@ -106,7 +120,7 @@ class b_plus_tree_map {
         return *this;
     }
 
-    b_plus_tree_map& operator=(b_plus_tree_map&& other) noexcept {
+    b_plus_tree_hash_set& operator=(b_plus_tree_hash_set&& other) noexcept {
         delete root_;
         root_ = other.root_;
         other.root_ = nullptr;
@@ -115,35 +129,29 @@ class b_plus_tree_map {
         return *this;
     }
 
-    ~b_plus_tree_map() {
+    ~b_plus_tree_hash_set() {
         delete root_;
         root_ = nullptr;
     }
 
-    [[nodiscard]] auto find(key_t key) noexcept {
+    [[nodiscard]] auto find(const T& value) {
         auto node = root_;
+        auto hash = HashT{}(value);
         while (!node->is_leaf()) {
-            node = node->as_inner()->find(key);
+            node = node->as_inner()->find(hash);
             if (node == nullptr) {
                 return end();
             }
         }
-        return node->as_leaf()->find(key);
+        return node->as_leaf()->find(hash, value);
     }
 
-    [[nodiscard]] auto find(key_t key) const noexcept {
-        return const_cast<b_plus_tree_map&>(*this).find(key);
+    [[nodiscard]] auto find(const T& value) const {
+        return const_cast<b_plus_tree_hash_set&>(*this).find(value);
     }
 
-    [[nodiscard]] auto lower_bound(key_t key) noexcept {
-        auto node = root_;
-        while (!node->is_leaf()) {
-            node = node->as_inner()->find(key);
-            if (node == nullptr) {
-                return end();
-            }
-        }
-        return node->as_leaf()->lower_bound_as_iter(key);
+    [[nodiscard]] size_t count(const T& value) const {
+        return const_cast<b_plus_tree_hash_set&>(*this).find(value) != end();
     }
 
     [[nodiscard]] auto begin() noexcept {
@@ -168,27 +176,46 @@ class b_plus_tree_map {
 
     template <typename... Args>
     auto emplace(Args&&... args) {
-        return try_emplace(std::forward<Args>(args)...);
+        T t(std::forward<Args>(args)...);
+        hash_t hash = HashT{}(t);
+        auto node = root_;
+        while (!node->is_leaf()) {
+            node = node->as_inner()->find_or_last(hash);
+        }
+        return node->as_leaf()->try_emplace(hash, *this, size_, std::move(t));
     }
 
     template <typename... Args>
-    auto try_emplace(key_t key, Args&&... args) {
-        auto node = root_;
-        while (!node->is_leaf()) {
-            node = node->as_inner()->find_or_last(key);
+    auto emplace_hint(const IterT& hint, Args&&... args) {
+        if (empty() || hint.is_end()) {
+            return emplace(std::forward<Args>(args)...).first;
         }
-        return node->as_leaf()->try_emplace(key, *this, size_, std::forward<Args>(args)...);
+        assert(hint.node_->is_leaf());
+
+        T t(std::forward<Args>(args)...);
+        auto hash = HashT{}(t);
+        auto node = hint.node_->as_leaf();
+
+        // The following may drop a valid hint but is easy to check.
+        if (node->data_.begin()->first > hash || (node->data_.end() - 1)->first < hash) {
+            return emplace(std::move(t)).first;
+        }
+
+        return node->try_emplace(hash, *this, size_, std::move(t)).first;
     }
 
-    void erase(key_t key) {
+    size_t erase(const T& value) {
         auto node = root_;
+        auto hash = HashT{}(value);
         while (!node->is_leaf()) {
-            node = node->as_inner()->find(key);
+            node = node->as_inner()->find(hash);
             if (node == nullptr) {
-                return;
+                return 0;
             }
         }
-        size_ -= node->as_leaf()->erase_key(key, *this);
+        auto n = node->as_leaf()->erase_key(hash, value, *this);
+        size_ -= n;
+        return n;
     }
 
     void erase(const IterT& iterator) {
@@ -201,10 +228,14 @@ class b_plus_tree_map {
         return size_;
     }
 
+    [[nodiscard]] bool empty() const noexcept {
+        return size_ == 0;
+    }
+
     void _check() {
         size_t count = 0;
         NLeafT* prev_leaf = nullptr;
-        key_t known_min = std::numeric_limits<key_t>::max();
+        hash_t known_min = std::numeric_limits<hash_t>::max();
         root_->_check(count, nullptr, prev_leaf, known_min, 0);
         assert(count == size());
     }
@@ -231,7 +262,7 @@ class b_plus_tree_map {
             return static_cast<NLeafT*>(this);
         }
 
-        virtual void _check(size_t&, NInnerT*, NLeafT*&, key_t&, key_t) = 0;
+        virtual void _check(size_t&, NInnerT*, NLeafT*&, hash_t&, hash_t) = 0;
 
       public:
         const bool is_leaf_;
@@ -243,15 +274,12 @@ class b_plus_tree_map {
         using DataIteratorT = decltype(std::vector<EntryT>().begin());
         friend IterT;
 
-        constexpr static size_t M_leaf = std::min(size_t(16), COUNT_MAX);
-        // Default MAX is 32. Special case for small COUNT with smaller inner leaf or
-        // trees with a single inner leaf. '*2' is added because leaf filling is not compact.
-        constexpr static size_t M_inner = std::min(size_t(16), COUNT_MAX / M_leaf * 2);
-        // TODO This could be improved but requires a code change to move > 1 entry when merging.
+        constexpr static size_t M_leaf = 16;
+        constexpr static size_t M_inner = 16;
+        // A value >2 requires a code change to move > 1 entry when merging.
         constexpr static size_t M_leaf_min = 2;   // std::max((size_t)2, M_leaf >> 2);
         constexpr static size_t M_inner_min = 2;  // std::max((size_t)2, M_inner >> 2);
-        // There is no point in allocating more leaf space than the max amount of entries.
-        constexpr static size_t M_leaf_init = std::min(size_t(8), COUNT_MAX);
+        constexpr static size_t M_leaf_init = 8;
         constexpr static size_t M_inner_init = 4;
 
       public:
@@ -274,10 +302,10 @@ class b_plus_tree_map {
             return this->is_leaf_ ? M_leaf_init : M_inner_init;
         }
 
-        [[nodiscard]] auto lower_bound(key_t key) noexcept {
+        [[nodiscard]] auto lower_bound(hash_t hash) noexcept {
             return std::lower_bound(
-                data_.begin(), data_.end(), key, [](EntryT& left, const key_t key) {
-                    return left.first < key;
+                data_.begin(), data_.end(), hash, [](EntryT& left, const hash_t hash) {
+                    return left.first < hash;
                 });
         }
 
@@ -287,7 +315,7 @@ class b_plus_tree_map {
 
         void erase_entry(DataIteratorT it_to_erase, TreeT& tree) {
             auto& parent_ = this->parent_;
-            key_t max_key_old = data_.back().first;
+            hash_t max_key_old = data_.back().first;
 
             data_.erase(it_to_erase);
             if (parent_ == nullptr) {
@@ -307,7 +335,7 @@ class b_plus_tree_map {
                 // Nothing to merge, just remove node. This should be rare, i.e. only happens when
                 // a rare 1-entry node has its last entry removed.
                 remove_from_siblings();
-                parent_->remove_node(max_key_old, tree);
+                parent_->remove_node(max_key_old, this, tree);
                 return;
             }
 
@@ -324,11 +352,11 @@ class b_plus_tree_map {
                         data_[0].second = nullptr;
                     }
                     auto prev_node = prev_node_;  // create copy because (this) will be deleted
-                    parent_->remove_node(max_key_old, tree);
+                    parent_->remove_node(max_key_old, this, tree);
                     if (prev_node->parent_ != nullptr) {
-                        key_t old1 = (prev_data.end() - 2)->first;
-                        key_t new1 = (prev_data.end() - 1)->first;
-                        prev_node->parent_->update_key(old1, new1);
+                        hash_t old1 = (prev_data.end() - 2)->first;
+                        hash_t new1 = (prev_data.end() - 1)->first;
+                        prev_node->parent_->update_key(old1, new1, prev_node);
                     }
                     return;
                 } else if (next_node_ != nullptr && next_node_->data_.size() < this->M_max()) {
@@ -341,34 +369,39 @@ class b_plus_tree_map {
                         next_data.emplace(next_data.begin(), std::move(data_[0]));
                         data_[0].second = nullptr;
                     }
-                    parent_->remove_node(max_key_old, tree);
+                    parent_->remove_node(max_key_old, this, tree);
                     return;
                 }
                 // This node is too small but there is nothing we can do.
             }
             if (it_to_erase == data_.end()) {
-                parent_->update_key(max_key_old, data_.back().first);
+                parent_->update_key(max_key_old, data_.back().first, this);
             }
         }
 
-        auto check_split(key_t key, TreeT& tree, DataIteratorT& it_in_out) {
+        struct SplitResult {
+            ThisT* node_;
+            DataIteratorT iter_;
+        };
+
+        SplitResult check_split(hash_t key, TreeT& tree, const DataIteratorT& it) {
             if (data_.size() < this->M_max()) {
                 if (this->parent_ != nullptr && key > data_.back().first) {
-                    this->parent_->update_key(data_.back().first, key);
+                    this->parent_->update_key(data_.back().first, key, this);
                 }
-                return static_cast<ThisT*>(this);
+                return {static_cast<ThisT*>(this), it};
             }
 
             ThisT* dest = this->split_node(key, tree);
             if (dest != this) {
                 // The insertion pos in node2 can be calculated:
-                auto old_pos = it_in_out - data_.begin();
-                it_in_out = dest->data_.begin() + old_pos - data_.size();
+                auto old_pos = it - data_.begin();
+                return {dest, dest->data_.begin() + old_pos - data_.size()};
             }
-            return dest;
+            return {dest, it};
         }
 
-        void _check_data(NInnerT* parent, key_t known_max) {
+        void _check_data(NInnerT* parent, hash_t known_max) {
             (void)parent;
             (void)known_max;
             // assert(parent_ == nullptr || data_.size() >= M_min);
@@ -381,7 +414,7 @@ class b_plus_tree_map {
         }
 
       private:
-        ThisT* split_node(key_t key, TreeT& tree) {
+        ThisT* split_node(hash_t key, TreeT& tree) {
             auto max_key = data_.back().first;
             if (this->parent_ == nullptr) {
                 auto* new_parent = new NInnerT(nullptr, nullptr, nullptr);
@@ -420,7 +453,7 @@ class b_plus_tree_map {
                 split_key = key;
             }
             this->parent_->update_key_and_add_node(
-                max_key, split_key, std::max(max_key, key), node2, tree);
+                max_key, split_key, std::max(max_key, key), this, node2, tree);
 
             // Return node for insertion of new value
             return key > split_key ? node2 : static_cast<ThisT*>(this);
@@ -435,7 +468,7 @@ class b_plus_tree_map {
             }
         }
 
-      protected:
+      public:
         std::vector<EntryT> data_;
         ThisT* prev_node_;
         ThisT* next_node_;
@@ -448,44 +481,57 @@ class b_plus_tree_map {
 
         ~bpt_node_leaf() noexcept = default;
 
-        [[nodiscard]] IterT find(key_t key) noexcept {
-            auto it = this->lower_bound(key);
-            if (it != this->data_.end() && it->first == key) {
-                return IterT(this, it);
+        [[nodiscard]] IterT find(hash_t hash, const T& value) noexcept {
+            PredT equals{};
+            IterT iter_full(this, this->lower_bound(hash));
+            while (!iter_full.is_end() && iter_full.hash() == hash) {
+                if (equals(*iter_full, value)) {
+                    return iter_full;
+                }
+                ++iter_full;
             }
             return IterT();
         }
 
-        [[nodiscard]] IterT lower_bound_as_iter(key_t key) noexcept {
-            auto it = this->lower_bound(key);
-            if (it != this->data_.end()) {
-                return IterT(this, it);
+        [[nodiscard]] auto lower_bound_value(hash_t hash, const T& value) noexcept {
+            PredT equals{};
+            IterT iter_full(this, this->lower_bound(hash));
+            while (!iter_full.is_end() && iter_full.hash() == hash) {
+                if (equals(*iter_full, value)) {
+                    break;
+                }
+                ++iter_full;
             }
-            return IterT();
+            return iter_full;
         }
 
-        template <typename... Args>
-        auto try_emplace(key_t key, TreeT& tree, size_t& entry_count, Args&&... args) {
-            auto it = this->lower_bound(key);
-            if (it != this->data_.end() && it->first == key) {
-                return std::make_pair(IterT(this, it), false);
+        auto try_emplace(hash_t hash, TreeT& tree, size_t& entry_count, T&& t) {
+            auto it = this->lower_bound(hash);
+            if (it != this->data_.end() && it->first == hash) {
+                // Hash collision !
+                PredT equals{};  // static?
+                IterT full_iter(this, it);
+                while (!full_iter.is_end() && full_iter.hash() == hash) {
+                    if (equals(*full_iter, t)) {
+                        return std::make_pair(full_iter, false);
+                    }
+                    ++full_iter;
+                }
             }
+            //            auto it = this->lower_bound_value(hash, t);
+            //            if (!it.is_end() && PredT{}(*it, t)) {
+            //                return std::make_pair(it, false);
+            //            }
             ++entry_count;
-
-            auto dest = this->check_split(key, tree, it);
-
-            auto x = dest->data_.emplace(
-                it,
-                std::piecewise_construct,
-                std::forward_as_tuple(key),
-                std::forward_as_tuple(std::forward<Args>(args)...));
-            return std::make_pair(IterT(this, x), true);
+            auto split_result = this->check_split(hash, tree, it);
+            auto it2 = split_result.node_->data_.emplace(split_result.iter_, hash, std::move(t));
+            return std::make_pair(IterT(split_result.node_, it2), true);
         }
 
-        bool erase_key(key_t key, TreeT& tree) {
-            auto it = this->lower_bound(key);
-            if (it != this->data_.end() && it->first == key) {
-                this->erase_entry(it, tree);
+        bool erase_key(hash_t hash, const T& value, TreeT& tree) {
+            auto iter = this->lower_bound_value(hash, value);
+            if (!iter.is_end() && PredT{}(*iter, value)) {
+                iter.node_->erase_entry(iter.iter_, tree);
                 return true;
             }
             return false;
@@ -496,12 +542,16 @@ class b_plus_tree_map {
         }
 
         void _check(
-            size_t& count, NInnerT* parent, NLeafT*& prev_leaf, key_t& known_min, key_t known_max) {
+            size_t& count,
+            NInnerT* parent,
+            NLeafT*& prev_leaf,
+            hash_t& known_min,
+            hash_t known_max) {
             this->_check_data(parent, known_max);
 
             assert(prev_leaf == this->prev_node_);
             for (auto& e : this->data_) {
-                assert(count == 0 || e.first > known_min);
+                assert(count == 0 || e.first >= known_min);
                 assert(this->parent_ == nullptr || e.first <= known_max);
                 ++count;
                 known_min = e.first;
@@ -523,29 +573,44 @@ class b_plus_tree_map {
             }
         }
 
-        [[nodiscard]] NodeT* find(key_t key) noexcept {
-            auto it = this->lower_bound(key);
+        [[nodiscard]] auto lower_bound_node(hash_t hash, const NodeT* node) noexcept {
+            auto it = this->lower_bound(hash);
+            while (it != this->data_.end() && it->first == hash) {
+                if (it->second == node) {
+                    return it;
+                }
+                ++it;
+            }
+            return this->data_.end();
+        }
+
+        [[nodiscard]] NodeT* find(hash_t hash) noexcept {
+            auto it = this->lower_bound(hash);
             return it != this->data_.end() ? it->second : nullptr;
         }
 
-        [[nodiscard]] NodeT* find_or_last(key_t key) noexcept {
-            auto it = this->lower_bound(key);
+        [[nodiscard]] NodeT* find_or_last(hash_t hash) noexcept {
+            auto it = this->lower_bound(hash);
             return it != this->data_.end() ? it->second : this->data_.back().second;
         }
 
-        void emplace_back(key_t key, NodeT* node) {
-            this->data_.emplace_back(key, node);
+        void emplace_back(hash_t hash, NodeT* node) {
+            this->data_.emplace_back(hash, node);
         }
 
         void _check(
-            size_t& count, NInnerT* parent, NLeafT*& prev_leaf, key_t& known_min, key_t known_max) {
+            size_t& count,
+            NInnerT* parent,
+            NLeafT*& prev_leaf,
+            hash_t& known_min,
+            hash_t known_max) {
             this->_check_data(parent, known_max);
 
             assert(this->parent_ == nullptr || known_max == this->data_.back().first);
             auto prev_key = this->data_[0].first;
             int n = 0;
             for (auto& e : this->data_) {
-                assert(n == 0 || e.first > prev_key);
+                assert(n == 0 || e.first >= prev_key);
                 e.second->_check(count, this, prev_leaf, known_min, e.first);
                 assert(this->parent_ == nullptr || e.first <= known_max);
                 prev_key = e.first;
@@ -553,14 +618,17 @@ class b_plus_tree_map {
             }
         }
 
-        void update_key(key_t old_key, key_t new_key) {
+        void update_key(hash_t old_key, hash_t new_key, NodeT* node) {
+            if (old_key == new_key) {
+                return;  // This can happen due to multiple entries with same hash.
+            }
             assert(new_key != old_key);
-            auto it = this->lower_bound(old_key);
+            auto it = this->lower_bound_node(old_key, node);
             assert(it != this->data_.end());
             assert(it->first == old_key);
             it->first = new_key;
             if (this->parent_ != nullptr && ++it == this->data_.end()) {
-                this->parent_->update_key(old_key, new_key);
+                this->parent_->update_key(old_key, new_key, this);
             }
         }
 
@@ -572,29 +640,41 @@ class b_plus_tree_map {
          * - Node1: key1_old > key1_new; Node 1 vs 2: key2 > new_key1
          */
         void update_key_and_add_node(
-            key_t key1_old, key_t key1_new, key_t key2, NodeT* child2, TreeT& tree) {
-            assert(key2 > key1_new);
+            hash_t key1_old,
+            hash_t key1_new,
+            hash_t key2,
+            NodeT* child1,
+            NodeT* child2,
+            TreeT& tree) {
+            // assert(key2 > key1_new);
             assert(key1_old >= key1_new);
-            auto it2 = this->lower_bound(key1_old) + 1;
+            auto it2 = this->lower_bound_node(key1_old, child1) + 1;
 
-            auto dest = this->check_split(key2, tree, it2);
+            auto split_result = this->check_split(key2, tree, it2);
             // check_split() guarantees that child2 is in the same node as child1
-            assert(it2 != dest->data_.begin());
+            assert(split_result.iter_ != split_result.node_->data_.begin());
             (it2 - 1)->first = key1_new;
-            child2->parent_ = dest;
-            dest->data_.emplace(it2, key2, child2);
+            child2->parent_ = split_result.node_;
+            child2->parent_->data_.emplace(it2, key2, child2);
         }
 
-        void remove_node(key_t key_remove, TreeT& tree) {
+        void remove_node(hash_t key_remove, NodeT* node, TreeT& tree) {
             auto it_to_erase = this->lower_bound(key_remove);
-            delete it_to_erase->second;
-            this->erase_entry(it_to_erase, tree);
+            while (it_to_erase != this->data_.end() && it_to_erase->first == key_remove) {
+                if (it_to_erase->second == node) {
+                    delete it_to_erase->second;
+                    this->erase_entry(it_to_erase, tree);
+                    return;
+                }
+                ++it_to_erase;
+            }
+            assert(false && "Node not found!");
         }
     };
 
     class bpt_iterator {
-        using EntryT = typename b_plus_tree_map<T, COUNT_MAX>::bpt_entry_leaf;
-        friend b_plus_tree_map<T, COUNT_MAX>;
+        using EntryT = typename b_plus_tree_hash_set<T, HashT, PredT>::bpt_entry_leaf;
+        friend b_plus_tree_hash_set<T, HashT, PredT>;
 
       public:
         using iterator_category = std::forward_iterator_tag;
@@ -604,7 +684,8 @@ class b_plus_tree_map {
         using reference = T&;
 
         // Arbitrary position iterator
-        explicit bpt_iterator(NLeafT* node, LeafIteratorT it) noexcept : node_{node}, iter_{it} {
+        explicit bpt_iterator(NLeafT* node, LeafIteratorT it) noexcept
+        : node_{it == node->data_.end() ? nullptr : node}, iter_{it} {
             assert(node->is_leaf_ && "just for consistency, insist that we iterate leaves only ");
         }
 
@@ -630,12 +711,12 @@ class b_plus_tree_map {
 
         auto& operator*() const noexcept {
             assert(AssertNotEnd());
-            return const_cast<EntryT&>(*iter_);
+            return const_cast<T&>(iter_->second);
         }
 
         auto* operator->() const noexcept {
             assert(AssertNotEnd());
-            return const_cast<EntryT*>(&*iter_);
+            return const_cast<T*>(&iter_->second);
         }
 
         auto& operator++() noexcept {
@@ -663,9 +744,18 @@ class b_plus_tree_map {
             return !(left == right);
         }
 
+        // TODO private
+        bool is_end() const noexcept {
+            return node_ == nullptr;
+        }
+
       private:
         [[nodiscard]] inline bool AssertNotEnd() const noexcept {
             return node_ != nullptr;
+        }
+
+        hash_t hash() {
+            return iter_->first;
         }
 
         NLeafT* node_;
@@ -676,6 +766,150 @@ class b_plus_tree_map {
     NodeT* root_;
     size_t size_;
 };
+
+template <
+    typename KeyT,
+    typename ValueT,
+    typename HashT = std::hash<KeyT>,
+    typename PredT = std::equal_to<KeyT>>
+class b_plus_tree_hash_map {
+    class iterator;
+    using IterT = iterator;
+    using EntryT = std::pair<KeyT, ValueT>;
+
+  public:
+    b_plus_tree_hash_map() : map_{} {};
+
+    b_plus_tree_hash_map(const b_plus_tree_hash_map&) = default;
+    b_plus_tree_hash_map(b_plus_tree_hash_map&&) noexcept = default;
+    b_plus_tree_hash_map& operator=(const b_plus_tree_hash_map&) = default;
+    b_plus_tree_hash_map& operator=(b_plus_tree_hash_map&&) noexcept = default;
+    ~b_plus_tree_hash_map() = default;
+
+    auto begin() const {
+        return IterT(map_.begin());
+    }
+
+    auto end() const {
+        return IterT(map_.end());
+    }
+
+    auto find(const KeyT& key) const {
+        return IterT(map_.find(EntryT{key, {}}));
+    }
+
+    auto count(const KeyT& key) const {
+        return map_.count(EntryT{key, {}});
+    }
+
+    template <typename... Args>
+    auto emplace(Args&&... args) {
+        return try_emplace(std::forward<Args>(args)...);
+    }
+
+    template <typename... Args>
+    auto emplace_hint(const IterT& hint, Args&&... args) {
+        return try_emplace(hint, std::forward<Args>(args)...);
+    }
+
+    template <typename... Args>
+    auto try_emplace(const KeyT& key, Args&&... args) {
+        auto result = map_.emplace(key, std::forward<Args>(args)...);
+        return std::make_pair(iterator(result.first), result.second);
+    }
+
+    template <typename... Args>
+    auto try_emplace(const IterT& hint, const KeyT& key, Args&&... args) {
+        auto result = map_.emplace_hint(hint.map_iter_, key, std::forward<Args>(args)...);
+        return iterator(result);
+    }
+
+    auto erase(const KeyT& key) {
+        return map_.erase({key, {}});
+    }
+
+    auto erase(const IterT& iterator) {
+        map_.erase(iterator.map_iter_);
+    }
+
+    auto size() const {
+        return map_.size();
+    }
+
+    auto empty() const {
+        return map_.empty();
+    }
+
+    void _check() {
+        map_._check();
+    }
+
+  private:
+    struct EntryHashT {
+        size_t operator()(const EntryT& x) const {
+            return HashT{}(x.first);
+        }
+    };
+
+    struct EntryEqualsT {
+        bool operator()(const EntryT& x, const EntryT& y) const {
+            return PredT{}(x.first, y.first);
+        }
+    };
+
+    class iterator {
+        using T = EntryT;
+        using MapIterType =
+            decltype(std::declval<b_plus_tree_hash_set<EntryT, EntryHashT, EntryEqualsT>>()
+                         .begin());
+        friend b_plus_tree_hash_map<KeyT, ValueT, HashT, PredT>;
+
+      public:
+        using iterator_category = std::forward_iterator_tag;
+        using value_type = T;
+        using difference_type = std::ptrdiff_t;
+        using pointer = T*;
+        using reference = T&;
+
+        explicit iterator(MapIterType map_iter) noexcept : map_iter_{map_iter} {}
+
+        // end() iterator
+        iterator() noexcept : map_iter_{} {}
+
+        auto& operator*() const noexcept {
+            return *map_iter_;
+        }
+
+        auto* operator->() const noexcept {
+            return &*map_iter_;
+        }
+
+        auto& operator++() noexcept {
+            ++map_iter_;
+            return *this;
+        }
+
+        auto operator++(int) noexcept {
+            IterT iterator(*this);
+            ++(*this);
+            return iterator;
+        }
+
+        friend bool operator==(const IterT& left, const IterT& right) noexcept {
+            return left.map_iter_ == right.map_iter_;
+        }
+
+        friend bool operator!=(const IterT& left, const IterT& right) noexcept {
+            return !(left == right);
+        }
+
+      private:
+        MapIterType map_iter_;
+    };
+
+    b_plus_tree_hash_set<EntryT, EntryHashT, EntryEqualsT> map_;
+};
+
 }  // namespace improbable::phtree
 
-#endif  // PHTREE_COMMON_B_PLUS_TREE_H
+#endif  // PHTREE_COMMON_B_PLUS_TREE_HASH_MAP_H