+++ phtree/phtree_multimap.h
@@ -17,6 +17,7 @@
 #ifndef PHTREE_PHTREE_MULTIMAP_H
 #define PHTREE_PHTREE_MULTIMAP_H
 
+#include "common/b_plus_tree_hash_map.h"
 #include "common/common.h"
 #include "v16/phtree_v16.h"
 #include <unordered_set>
@@ -186,7 +187,7 @@ template <
     dimension_t DIM,
     typename T,
     typename CONVERTER = ConverterNoOp<DIM, scalar_64_t>,
-    typename BUCKET = std::unordered_set<T>,
+    typename BUCKET = b_plus_tree_hash_set<T>,
     bool POINT_KEYS = true,
     typename DEFAULT_QUERY_TYPE = QueryPoint>
 class PhTreeMultiMap {
@@ -693,14 +694,14 @@ template <
     dimension_t DIM,
     typename T,
     typename CONVERTER = ConverterIEEE<DIM>,
-    typename BUCKET = std::unordered_set<T>>
+    typename BUCKET = b_plus_tree_hash_set<T>>
 using PhTreeMultiMapD = PhTreeMultiMap<DIM, T, CONVERTER, BUCKET>;
 
 template <
     dimension_t DIM,
     typename T,
     typename CONVERTER_BOX,
-    typename BUCKET = std::unordered_set<T>>
+    typename BUCKET = b_plus_tree_hash_set<T>>
 using PhTreeMultiMapBox = PhTreeMultiMap<DIM, T, CONVERTER_BOX, BUCKET, false, QueryIntersect>;
 
 /**
@@ -713,7 +714,7 @@ template <
     dimension_t DIM,
     typename T,
     typename CONVERTER_BOX = ConverterBoxIEEE<DIM>,
-    typename BUCKET = std::unordered_set<T>>
+    typename BUCKET = b_plus_tree_hash_set<T>>
 using PhTreeMultiMapBoxD = PhTreeMultiMapBox<DIM, T, CONVERTER_BOX, BUCKET>;
 
 }  // namespace improbable::phtree